modify quam + quam on the fly
